Edit login screen view, and localized it
