+++ Resources/Assets.xcassets/clic&farm/clic&farm-logo.imageset/Contents.json
@@ -2,11 +2,11 @@
   "images" : [
     {
       "idiom" : "universal",
+      "filename" : "logo-splash-screen.png",
       "scale" : "1x"
     },
     {
       "idiom" : "universal",
-      "filename" : "Icon-180.png",
       "scale" : "2x"
     },
     {
